Adjusted files to be more organized. (Maybe I don't need any routes we will see.)
