+++ src/components/homePage/homePage.tsx
@@ -1,9 +1,11 @@
 import { Button } from '@mui/material';
+import NavBars from '../navBars/navBars';
 
 export default function HomePage() {
     return (
         <div>
            Portfolio
+           <NavBars />
         </div>
     )
 }
